MAke DockerRegistryClient a facade in front of what will become v1 and v2 API classes
